+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 3.0.0-SNAPSHOT
 ### Changed
+- [#281](https://github.com/krux/hyperion/issues/281) - Support for not failing on un-defined pipeline parameters
 - [#291](https://github.com/krux/hyperion/issues/291) - Clean up the implicits
 - [#285](https://github.com/krux/hyperion/issues/285) - SnsAlarm requires topic arn and added default subject and message
 - [#286](https://github.com/krux/hyperion/issues/286) - Fix a bug in 3.0 that main class in jar activity is incorrect